+++ App/src/commands/context-menu/avatar.js
@@ -1,0 +1,10 @@
+const { ContextMenuCommandBuilder, ApplicationCommandType } = require('discord.js');
+
+module.exports = {
+    data: new ContextMenuCommandBuilder()
+        .setName('avatar')
+        .setType(ApplicationCommandType.User),
+    async execute(interaction) {
+        return interaction.reply(`${interaction.targetUser.username}'s [avatar](${interaction.targetUser.displayAvatarURL({ dynamic: true, size: 512 })})`);
+    },
+};