+++ App/src/commands/utility/bot-check.js
@@ -11,10 +11,20 @@ module.exports = {
         ),
     async execute(interaction) {
         const user = interaction.options.getUser('user') || interaction.user;
-        const embed = new EmbedBuilder()
-            .setAuthor({ name: `${user.tag} botcheck ${user.bot}!`, iconURL: `${user.displayAvatarURL({ dynamic: true, size: 512 })}` })
-            .setColor(0xFFFFFF);
-        const message = await interaction.reply({ embeds: [embed], fetchReply: true });
-        message.react('🤖');
+        let embed, message;
+        if (user.bot) {
+            embed = new EmbedBuilder()
+                .setAuthor({ name: `${user.tag} is a BOT!`, iconURL: `${user.displayAvatarURL({ dynamic: true, size: 512 })}` })
+                .setColor(0xFFFFFF);
+            message = await interaction.reply({ embeds: [embed], fetchReply: true });
+            message.react('🤖');
+        }
+        else {
+            embed = new EmbedBuilder()
+                .setAuthor({ name: `${user.tag} is not a BOT.`, iconURL: `${user.displayAvatarURL({ dynamic: true, size: 512 })}` })
+                .setColor(0xFFFFFF);
+            message = await interaction.reply({ embeds: [embed], fetchReply: true });
+            message.react('👤');
+        }
     },
 };