+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: 6d37ba7 - v102
+| Author: FJrodafo - Sat, 6 Jul 2024 14:25:40 +0100
+| 
 * Commit: 1e75b49 - v101
 | Author: FJrodafo - Sat, 6 Jul 2024 11:06:18 +0100
 |